moved set eval funcs to eval.go
